# reference/IrochiCy-main/IrochiCy-main/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SIH26145 API",
    "description": "Network Threat Detection \u2014 Security Intelligence Hub",
    "version": "1.0.0"
  },
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "description": "Authenticate and receive access + refresh tokens.",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Logout",
        "description": "Log out the current user (audit trail only \u2014 JWT is stateless).",
        "operationId": "logout_auth_logout_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Logout Auth Logout Post"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/alerts/": {
      "get": {
        "tags": [
          "alerts"
        ],
        "summary": "List Alerts",
        "description": "List alerts with filtering and pagination.",
        "operationId": "list_alerts_alerts__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "threat_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Threat Type"
            }
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "src_ip",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Src Ip"
            }
          },
          {
            "name": "dst_ip",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dst Ip"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/alerts/export/csv": {
      "get": {
        "tags": [
          "alerts"
        ],
        "summary": "Export Alerts Csv",
        "description": "Export alerts as a streaming CSV download.",
        "operationId": "export_alerts_csv_alerts_export_csv_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "threat_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Threat Type"
            }
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/dashboard/summary": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Get Summary",
        "description": "Full dashboard summary (5-second Redis cache).",
        "operationId": "get_summary_dashboard_summary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardSummary"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/dashboard/kpi": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Get Kpi",
        "description": "Live KPI metrics.",
        "operationId": "get_kpi_dashboard_kpi_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KPIData"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/dashboard/timeline": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Get Timeline",
        "description": "24-hour alert timeline bucketed by threat type.",
        "operationId": "get_timeline_dashboard_timeline_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/threats/": {
      "get": {
        "tags": [
          "threats"
        ],
        "summary": "List Threat Types",
        "description": "List all 5 threat type stats with live DB counts and detector status.",
        "operationId": "list_threat_types_threats__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ThreatTypeStat"
                  },
                  "type": "array",
                  "title": "Response List Threat Types Threats  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/users/me": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get Me",
        "description": "Get the currently authenticated user's profile.",
        "operationId": "get_me_users_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "users"
        ],
        "summary": "Update Me",
        "description": "Update the current user's own profile (name and email only for analysts).",
        "operationId": "update_me_users_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AlertListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AlertResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "has_next": {
            "type": "boolean",
            "title": "Has Next"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "has_next"
        ],
        "title": "AlertListResponse"
      },
      "AlertResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "threat_type": {
            "type": "string",
            "title": "Threat Type"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "src_ip": {
            "type": "string",
            "title": "Src Ip"
          },
          "dst_ip": {
            "type": "string",
            "title": "Dst Ip"
          },
          "src_port": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Src Port"
          },
          "dst_port": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dst Port"
          },
          "protocol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Protocol"
          },
          "detector_id": {
            "type": "string",
            "title": "Detector Id"
          },
          "schema_version": {
            "type": "string",
            "title": "Schema Version"
          },
          "sensor_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sensor Source"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceItem"
            },
            "type": "array",
            "title": "Evidence"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "assigned_to": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned To"
          },
          "analyst_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analyst Notes"
          },
          "raw_event_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Event Id"
          },
          "ingest_latency_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ingest Latency Ms"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "threat_type",
          "severity",
          "confidence",
          "src_ip",
          "dst_ip",
          "src_port",
          "dst_port",
          "protocol",
          "detector_id",
          "schema_version",
          "sensor_source",
          "evidence",
          "status",
          "assigned_to",
          "analyst_notes",
          "raw_event_id",
          "ingest_latency_ms",
          "created_at",
          "updated_at"
        ],
        "title": "AlertResponse"
      },
...
      "DashboardSummary": {
        "properties": {
          "kpi": {
            "$ref": "#/components/schemas/KPIData"
          },
          "top_threats": {
            "items": {
              "$ref": "#/components/schemas/TopThreat"
            },
            "type": "array",
            "title": "Top Threats"
          },
          "top_source_ips": {
            "items": {
              "$ref": "#/components/schemas/TopSourceIP"
            },
            "type": "array",
            "title": "Top Source Ips"
          }
        },
        "type": "object",
        "required": [
          "kpi",
          "top_threats",
          "top_source_ips"
        ],
        "title": "DashboardSummary"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "KPIData": {
        "properties": {
          "total_alerts_today": {
            "type": "integer",
            "title": "Total Alerts Today"
          },
          "active_threats": {
            "type": "integer",
            "title": "Active Threats"
          },
          "events_per_second": {
            "type": "number",
            "title": "Events Per Second"
          },
          "detectors_active": {
            "type": "integer",
            "title": "Detectors Active"
          },
          "pipeline_latency_p95_ms": {
            "type": "integer",
            "title": "Pipeline Latency P95 Ms"
          },
          "alerts_delta_today": {
            "type": "integer",
            "title": "Alerts Delta Today"
          }
        },
        "type": "object",
        "required": [
          "total_alerts_today",
          "active_threats",
          "events_per_second",
          "detectors_active",
          "pipeline_latency_p95_ms",
          "alerts_delta_today"
        ],
        "title": "KPIData"
      },
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Username"
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest"
      },
...
      "ThreatTypeStat": {
        "properties": {
          "threat_type": {
            "type": "string",
            "title": "Threat Type"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "signals": {
            "items": {
              "$ref": "#/components/schemas/SignalDefinition"
            },
            "type": "array",
            "title": "Signals"
          },
          "alerts_today": {
            "type": "integer",
            "title": "Alerts Today"
          },
          "alerts_7d": {
            "type": "integer",
            "title": "Alerts 7D"
          },
          "avg_confidence": {
            "type": "number",
            "title": "Avg Confidence"
          },
          "detector_status": {
            "type": "string",
            "enum": [
              "running",
              "idle",
              "error"
            ],
            "title": "Detector Status"
          },
          "last_detection": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Detection"
          }
        },
        "type": "object",
        "required": [
          "threat_type",
          "display_name",
          "description",
          "signals",
          "alerts_today",
          "alerts_7d",
          "avg_confidence",
          "detector_status",
          "last_detection"
        ],
        "title": "ThreatTypeStat"
      },
...
      "TimelineResponse": {
        "properties": {
          "buckets": {
            "items": {
              "$ref": "#/components/schemas/TimelineBucket"
            },
            "type": "array",
            "title": "Buckets"
          }
        },
        "type": "object",
        "required": [
          "buckets"
        ],
        "title": "TimelineResponse"
      },
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token"
        ],
        "title": "TokenResponse"
      },
      "TopSourceIP": {
        "properties": {
          "ip": {
            "type": "string",
            "title": "Ip"
          },
          "event_count": {
            "type": "integer",
            "title": "Event Count"
          },
          "alert_count": {
            "type": "integer",
            "title": "Alert Count"
          },
          "last_seen": {
            "type": "string",
            "format": "date-time",
            "title": "Last Seen"
          },
          "top_threat": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top Threat"
          }
        },
        "type": "object",
        "required": [
          "ip",
          "event_count",
          "alert_count",
          "last_seen",
          "top_threat"
        ],
        "title": "TopSourceIP"
      },
      "TopThreat": {
        "properties": {
          "threat_type": {
            "type": "string",
            "title": "Threat Type"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "percentage": {
            "type": "number",
            "title": "Percentage"
          }
        },
        "type": "object",
        "required": [
          "threat_type",
          "count",
          "percentage"
        ],
        "title": "TopThreat"
      },
...
      "UserResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "last_login_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "full_name",
          "email",
          "role",
          "is_active",
          "last_login_at",
          "created_at"
        ],
        "title": "UserResponse"
      },
      "UserUpdate": {
        "properties": {
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "role": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(admin|analyst)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          }
        },
        "type": "object",
        "title": "UserUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        }
      }
    }
  }
}